Clean up imports and docs for can reader
