+++ can/CAN.py
@@ -132,16 +132,18 @@ class LogReader(object):
     Replay logged CAN messages from a file.
 
     The format is determined from the file format which can be one of:
-      * .asc: :class:`can.ASCWriter`
-      * .csv: :class:`can.CSVWriter`
-      * .db: :class:`can.SqliteWriter`
+      * .asc
+      * .csv
+      * .db
 
-    Options:
+    Exposes a simple iterator interface, to use simply:
 
-    - Ignore timestamps
-    - Minimum time between messages
-    - Skip periods of inactivity greater than X
+        >>> for m in LogReader(my_file):
+        ...     print(m)
 
+    Note there are no time delays, if you want to reproduce
+    the measured delays between messages look at the
+    :class:`can.util.MessageSync` class.
     """
 
     @classmethod
@@ -153,51 +155,12 @@ class LogReader(object):
             raise NotImplemented
         #     return CSVReader(filename)
         if filename.endswith(".db"):
-            return SqliteReader(filename)
+            return SqlReader(filename)
 
 
-class MessageSync:
-
-    def __init__(self, messages, timestamps=True, gap=0.0001, skip=60):
-        """
-
-        :param messages: An iterable of :class:`can.Message` instances.
-        :param timestamps: Use the messages' timestamps.
-        :param gap: Minimum time between sent messages
-        :param skip: Skip periods of inactivity greater than this.
-        """
-        self.raw_messages = messages
-        self.timestamps = timestamps
-        self.gap = gap
-        self.skip = skip
-
-    def __iter__(self):
-        log.debug("Iterating over messages at real speed")
-        playback_start_time = time.time()
-        recorded_start_time = None
-
-        for m in self.raw_messages:
-            if recorded_start_time is None:
-                recorded_start_time = m.timestamp
-
-            if self.timestamps:
-                # Work out the correct wait time
-                now = time.time()
-                current_offset = now - playback_start_time
-                recorded_offset_from_start = m.timestamp - recorded_start_time
-                remaining_gap = recorded_offset_from_start - current_offset
-
-                sleep_period = max(self.gap, min(self.skip, remaining_gap))
-            else:
-                sleep_period = self.gap
-
-            time.sleep(sleep_period)
-            yield m
-
-
-class SqliteReader:
+class SqlReader:
     def __init__(self, filename):
-        log.debug("Starting sqlitereader with {}".format(filename))
+        log.debug("Starting sqlreader with {}".format(filename))
         conn = sqlite3.connect(filename)
 
         self.c = conn.cursor()
@@ -217,7 +180,7 @@ class SqliteReader:
     def __iter__(self):
         log.debug("Iterating through messages from sql db")
         for frame_data in self.c.execute("SELECT * FROM messages"):
-            yield SqliteReader.create_frame_from_db_tuple(frame_data)
+            yield SqlReader.create_frame_from_db_tuple(frame_data)
 
 
 class Printer(Listener):
